requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# HTTP & Scraping
httpx==0.27.2
aiohttp==3.10.8
beautifulsoup4==4.12.3
lxml==5.3.0

# AI
anthropic==0.34.2

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Gmail Integration
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# Task Queue (for background scraping)
celery[redis]==5.4.0
redis==5.1.1

# Utilities
tenacity==9.0.0
